+++ app/Database/Migrations/2025-05-29-075616_AddImageSizesToFilledFiles.php
@@ -19,6 +19,12 @@ class AddImageSizesToFilledFiles extends Migration
 
     public function down()
     {
+    $db = \Config\Database::connect();
+
+        if ($db->DBDriver === 'SQLite3') {
+            return;
+        }
+
         $this->forge->dropColumn('filledFiles', 'imageSizes');
     }
 }
